requirements.txt
beautifulsoup4==4.12.3
python-telegram-bot==21.1.1
requests==2.32.3
feedparser==6.0.11
python-dotenv==1.0.1